+++ tests/spec/core_functions/meta/load_css/extend.rs
@@ -8,6 +8,19 @@ fn runner() -> crate::TestRunner {
         .mock_file("in_input/before/_other.scss", "a {b: c}\n")
         .mock_file("in_other/after/_other.scss", "d {@extend a !optional}\n")
         .mock_file("in_other/before/_other.scss", "d {@extend a !optional}\n")
+        .mock_file(
+            "shared_cssless_midstream/_midstream.scss",
+            "@use 'upstream';\n",
+        )
+        .mock_file(
+            "shared_cssless_midstream/_target.scss",
+            "@use 'midstream';\n\n.target {a: b}\n",
+        )
+        .mock_file("shared_cssless_midstream/_upstream.scss", "@c;\n")
+        .mock_file(
+            "shared_cssless_midstream/extender.scss",
+            "@use 'target';\n\n.extender {\n  @extend .target;\n}\n",
+        )
 }
 
 mod in_input {
@@ -75,4 +88,23 @@ mod in_other {
          \n}\n"
         );
     }
+}
+#[test]
+#[ignore] // wrong result
+fn shared_cssless_midstream() {
+    let runner = runner().with_cwd("shared_cssless_midstream");
+    assert_eq!(
+        runner.ok("// Regression test for sass/sass#3322\
+             \n@use \'sass:meta\';\
+             \n@use \'target\';\n\
+             \n@include meta.load-css(\'extender\');\n"),
+        "@c;\
+         \n.target {\
+         \n  a: b;\
+         \n}\
+         \n@c;\
+         \n.target, .extender {\
+         \n  a: b;\
+         \n}\n"
+    );
 }